+++ src/app/App.module.css
@@ -1,3 +1,11 @@
+.home-page {
+  min-height: calc(100dvh - 57px);
+}
+
+.card {
+  width: 1024px;
+}
+
 .loading-page {
   min-height: calc(100dvh - 57px);
 }
