--- src/app/posts/Posts.module.css
@@ -1,7 +1,0 @@
-.posts-page {
-  min-height: calc(100dvh - 57px);
-}
-
-.card {
-  width: 1024px;
-}
